refactor: split webRTCView Header
