+++ src/Components/WebRTCView/Header/Header.const.ts
@@ -1,0 +1,12 @@
+export const AUCTION_HEADER_IMG = [
+  {
+    id: 1,
+    src: '/asset/Auction/Sound.svg',
+    alt: '음소거',
+  },
+  {
+    id: 2,
+    src: '/asset/Auction/Exit.svg',
+    alt: '나가기',
+  },
+];
