+++ src/Components/WebRTCView/Header/Header.style.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const HeaderImgContainer = styled.div`
+  img {
+    cursor: pointer;
+    width: 30px;
+    height: 30px;
+    margin-right: 15px;
+  }
+`;
+
+export default { HeaderImgContainer };
